# openapi.yaml
...
openapi: 3.1.0
info:
  title: Открывай API
  description: REST API мини-приложения «Открывай» в MAX. Все методы /api требуют
    заголовок X-Max-Init-Data — строку WebApp.initData из MAX Bridge; подпись
    проверяется по токену бота (HMAC-SHA256).
  version: 0.1.0
components:
  securitySchemes:
    maxInitData:
      type: apiKey
      in: header
      name: x-max-init-data
      description: WebApp.initData из MAX Bridge
    maxWebhookSecret:
      type: apiKey
      in: header
      name: x-max-bot-api-secret
      description: Секрет подписки, который MAX передаёт при доставке обновлений
...
paths:
...
  /webhook/max:
    post:
      operationId: maxWebhook
      summary: Приём обновлений MAX Bot API (только при BOT_MODE=webhook)
      tags:
        - bot
      security:
        - maxWebhookSecret: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    enum:
                      - true
                required:
                  - ok
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
...
servers:
  - url: https://otkryvay.stasvinokur.ru
    description: Публичный стенд